% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foydle.R
\name{foydle}
\alias{foydle}
\title{Fast Pearson Correlation For Linear Regression Interaction Term}
\usage{
foydle(xmat, ymat, zmat, output_file = NULL, colnames = c("x", "y", "z"),
  pvalue_threshold = NULL, cores = 1)
}
\arguments{
\item{xmat, ymat, zmat}{Numeric matrices}

\item{output_file}{Path to output file}

\item{colnames}{Names for columns 1-3 in output table}

\item{pvalue_threshold}{Save only results with p-value below
threshold (default: save all results)}

\item{cores}{Number of CPU cores to use (default 1)}
}
\value{
A data frame with columns "x", "y", "z", and "r".  Columns
    "x", "y", and "z" contain the names of the columns of
    \code{xmat}, \code{ymat}, and \code{zmat} that were used in
    computing the corresponding Pearson correlation coefficient in
    the "r" column.  The names of columns other than the r-column
    can be changed via the \code{colnames} argument.  The same data
    frame is written to \code{output_file}.
}
\description{
\code{foydle} computes Pearson correlation coefficients for the
interaction term in the linear regression
    \deqn{x = \beta_0 + \beta_1 y + \beta_2 z + \beta_3 y*z}{x = b0 + b1 y + b2 z + b3 y*z}
where \eqn{x}, \eqn{y}, and \eqn{z} are columns from \code{xmat},
\code{ymat}, and \code{zmat}, respectively.
}

