% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foydle.R
\name{foydle}
\alias{foydle}
\title{Fast Pearson correlation for linear regression interaction term}
\usage{
foydle(xmat, ymat, zmat, output_file)
}
\arguments{
\item{xmat, ymat, zmat}{Numeric matrices}

\item{output_file}{Path to output file}
}
\value{
Return \code{output_file} where a data frame with columns
    "x", "y", "z", and "r" was written.  Columns "x", "y", and "z"
    contain the names of the columns of \code{xmat}, \code{ymat},
    and \code{zmat} that were used in computing the corresponding
    Pearson correlation coefficient in the "r" column.
}
\description{
\code{foydle} computes Pearson correlation coefficients for the
interaction term in the linear regression
}
\details{
\deqn{x = 1 + y + z + y*z}

where \eqn{x}, \eqn{y}, and \eqn{z} are columns from \code{xmat},
\code{ymat}, and \code{zmat}, respectively.
}

